+++ test/browserstack.js
@@ -48,7 +48,8 @@ define([
     // { browser: 'Safari', browser_version: '7.1', os: 'OS X', os_version: 'Mavericks', platform: 'MAC', browserName: 'Safari 7' },
     // { browser: 'Safari', browser_version: '6.2', os: 'OS X', os_version: 'Mountain Lion', platform: 'MAC', browserName: 'Safari 6' },
 
-    { browserName: 'iPhone', platform: 'MAC', device: 'iPhone 6S' },
+    // Disabled because tests are flaky and BrowserStack-Support claims Intern is the culprit
+    // { browserName: 'iPhone', platform: 'MAC', device: 'iPhone 6S' },
   ];
   /*eslint-enable camelcase */